requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-bootstrap-v5==1.0.11
django-resized==1.0.2
Pillow==10.1.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
asgiref==3.7.2
beautifulsoup4==4.12.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-bootstrap-v5==1.0.11
django-import-export==3.3.2
django-resized==1.0.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
